Remove background tasks
@johnezang and @appleguy brought up issues with background tasks and
performance (specifically in our app we were seeing the creation of the
tasks take a significant portion of all CPU time on single core devices).

Background tasks are created for the following actions:

- Emptying the trash
- Setting an object
- Removing an object
- Trimming the cache (size, date, size/date)
- Removing all objects
- Enumerating all objects

In the above list, only one task is truly necessary to complete to
avoid getting in an inconsistent state in a subsequent run, setObject:

After reasoning about it for  some time, I believe that creating
the background tasks is no longer necessary because of our try catch
on object unarchive.

The background tasks mainly ensure the cache stays in a consistent state
however they cannot guarantee this as an app can crash or be killed at
any time.

This also has the added advantage of simplifying app extensions which
never supported this in the first place.