[PINDiskCache] Respect small byteLimit settings by checking object size in setObject:
This serves as a mechanism for some clients, like users of the Texture framework, to
effectively disable the PINDiskCache by setting a byteLimit of 1.

Before this change, the cache would transiently exceed the byte limit by writing
the file anyway, and then immediately deleting it.

This change will not affect most users of PINDiskCache, but is important for this
specific use case.