Asserts are enabled in release when using Swift Package Manager
When using Swift Package Manager to integrate `PINCache` into an app, Xcode does not automatically set `NS_BLOCK_ASSERTIONS` to `1` when building the package in release. This means that any `NSAsserts` that are hit in release cause crashes. This appears to only be an issue with ObjC packages. https://forums.swift.org/t/assertions-in-swift-packages/42692

The change I've made is to `Package.swift` to set `NS_BLOCK_ASSERTIONS` to `1` when building in release.