Add ability to configure maxConcurrentOperations
## Summary

PINCache uses a PINOperationQueue to perform async operations on a background queue, while preventing thread explosion. However, the amount is hardcoded to 10 and I'd like to be able to experiment with how much concurrency we use. Therefore, add a property to allow it to be configuration.

## Testing

Since it's just a configuration parameter, make sure the unit tests still pass.